Add basic support for diffs in HTML format.
Diff.pm now supports a 'mime' query parameter, that corresponds to the
mimetype of the output.  The default is 'text/html'.  'text/plain' is
also supported.

When emitting HTML, use the new 'diff' template, using Text::Diff::HTML
to generate the actual HTML.  This can be styled by the new diff.css,
but is not yet done so by default.