Change how logs are processed.
Previously, any file copies were traversed when the log was being
retrieved.  This meant lots of log output, which is good.

However, the "Show diffs" and "Checkout this file at rev x"
functionality was broken for any output for files that had
been copied.  This is because the links all referred to the
file's new name, even when the log output was for the file's
old name.

Now, log processing stops at a copy point.  However, the log
output has been augmented, and includes a list of all the
files affected by the change (based on code from Revisions.pm).

This has the side effect that the log entry for a copy now includes 
the information about where the file was copied from.

Using some of the code from the Template/revision, Template/log
can now display a link to the source of the copy, allowing the
reader to easily jump to the source of the copy, and continue
reading the log.