The navigation bar shows you the path from the root of the repository down to wherever you're viewing.  Each component is a link.
The old behaviour was that the link's action would change depending on
what you're doing.  So if you're browsing, and click on one of the links
you'll go on browsing further up the tree.

But if you were viewing the log, or looking at a specific revision,
then clicking the link would take you up the tree but stay with the
current action.

I've found this confusing, so disable it.  Now the navigation bar links
always take you to browse wherever you've clicked on.