Add a new command, 'view', with associated module and template.
This command takes a repository path and an (optional) revision.  If
the revision is not specified then it defaults to the repository's
youngest revision.

It searches the history to find the youngest interesting revision from
the one given.  It then displayes the log message for that revision, 
and the contents of the file at that revision.

The effect is similar to Trac's default behaviour for viewing a file.
For a Trac  example, see 

 http://jc.ngo.org.uk/trac-bin/trac.cgi/file/nik/libtap/trunk/src/tap.3?rev=353