Change the browse behaviour.
Files will now appear as links to the 'view' command, not the 'log'
command.  If you click on a file you're more likely to want to see
the contents of that file than it's log history.

Revision numbers are now links to the 'revision' command, showing
you all the changes for that revision.