There's a problem with the "Revision Log" link.
Say you're viewing a file at rev 2.  The repo current version is 5, and
the file was removed from the repo in rev 3.

The old link would try and view the log for the file starting from rev 5
(the repo's youngest rev).  This would generate an error, as the file
doesn't exist in the repo at that revision.

The new link starts viewing the log from whatever revision of the file
you were view'ing.  This is guaranteed to work, but might be slightly
confusing.  Needs user feedback to see if this is the right fix.