Add detailed documentation about configuring SVN::Web via config.yaml.
For consistency, rename @PLUGINS to @DEFAULT_ACTIONS.  This list can now
be overridden within config.yaml, and the action passed to the CGI script
is checked to make sure that it's allowed.

This makes it possible to create a restricted SVN::Web installation -- for
example, the 'checkout' action could be removed from the configuration file.
This also makes it easier to add new actions to the list.

Add a "meta-filter" for log messages.  This is a Template::Toolkit filter
that can be configured to run multiple filters from a single filter
directive in the Template.  If the meta-filter is configured to run any
filters that don't exist on the system then they are silently skipped.

This means that the filtering of log messages can be controlled from the
config file, rather than requiring a change to every template that displays
a log message.  It also means that we can ship with several useful filters
turned on (eg, ::Clickable, and Clickable::Email) and still work if the
user hasn't installed those filters.

Adjust the templates that display log messages to use this new filter.  
Change the config file installed by svnweb-install to specify a sensible
default set.

There's quite a lot of my work gone in here, so add myself to the AUTHORS
and COPYRIGHT sections.