Try and deal with revision number 'stickiness'.
Revision numbers are sticky.  Once you start browsing a particular revision
then you're stuck there.  Every browse, view, log, and so on is tied to that
revision.

This is so that a link isn't inadvertently generated to a file that no longer
exists in that revision.

This is fine, except the default behaviour is to almost immediately lock the
user to a revision.  This can be confusing.  Change the template behaviour.

1.  The browse, log, and view actions return an additional 'youngest_rev'
    parameter to the templates.  This is either the repo's youngest
    revision, or, in the case of 'view', it's the file's youngest
    interesting revision.

2.  The header now displays the revision number that you're viewing, or
    'HEAD' if you're looking at the most recent revision, to make it more
    obvious to the viewer whether or not they're stuck on a revision.

3.  Any URLs where rev == youngest_rev are generated without a trailing
    rev=... parameter.  This makes it much more likely that, when browsing,
    the user will stay at the HEAD of the repo.