Turns out you can't ask SVN for a node's history from the repo's youngest rev if the node no longer exists in the youngest rev.  There's no way of saying "Given this path in the repo, when was it deleted?".
So use a different heuristic in the view action.  If the URI had no rev=
param then we know they wanted the file's HEAD.  In which case youngest_rev
is the same as the file's most recent interesting revision.  In other cases,
use the fs' youngest rev as the youngest_rev template parameter.