Implement errors-as-exceptions.
List Exception::Class as a prerequisite in Makefile.PL.

Add SVN::Web::X, which derives from it, and provides a documented interface
to throw exceptions.

Adjust all the actions that used to die() to throw exceptions instead.
Ditto in SVN::Web.  Have SVN::Web catch exceptions and send them to an
'x' template for formatting.

Include an 'x' template for the trac/ and plain/ templates.

Provide friendly messages for each exception in en.po.