Change how the actions: config directive is processed.  Instead of a list of actions with separate <action>_class directives, turn it in to a hash of hashes.  The primary key is the action name, with "class" subkeys that specify the class that implements the action.
Update the documentation for each action to reflect the new syntax, and
the docs in Web.pm.  Note the change in UPDATING.pod, and have svnweb-install
include the directive in the config.yaml that it writes.

Remove some 'branch' related code in Web.pm and Log.pm that got in the way,
but is unrelated to the other changes.