Add svnweb-server, to make it easier for new users to get started with SVN::Web.  Adjust Makefile.PL and MANIFEST accordingly.
Update the documentation in Web.pm, be clearer about the steps necessary
to get started, and include more detailed documentation about configuring
Apache with/without mod_perl to use SVN::Web.

Bring CHANGES.pod up to date with changes for this release.

Bump the version number to 0.43 in preparation for a release.

Add t/3svnweb-install.t to MANIFEST.